fix: reorganize table files
